backend/recommendation/requirements.txt
attrs==19.3.0
numpy==1.20.1
scipy==1.4.1
jsonpickle==1.4.1
redis==3.5.3